refactor!: move shared to data
