+++ bridge/qb/server/main.lua
@@ -4,7 +4,7 @@ require 'bridge.qb.server.debug'
 require 'bridge.qb.server.events'
 
 local convertItems = require 'bridge.qb.shared.compat'.convertItems
-convertItems(require '@ox_inventory.data.items', require 'shared.items')
+convertItems(require '@ox_inventory.data.items', require 'data.items')
 
 qbCoreCompat = {}
